Update CI python version
Because python 3.6 has reached EOL and was dropped from the
ubuntu-latest image (22.04), the CI will fail when setting up the
environment.
See https://github.com/kmkurn/pytorch-crf/issues/117

This updates the python version in the CI matrix to use the oldest
supported minor python verion.
The torch version is also updated to match the available build on PyPI.
Testing on oldest torch 2.x is also added.